requirements.txt
django
djangorestframework
django-cors-headers
Pillow
graphviz
pydotplus
djangorestframework-simplejwt
django-import-export


